Set up Post blade component with its own view logic
